hard shit

+++ 0x00-hello_world/6-size.c
@@ -10,12 +10,11 @@ int main(void)
 	long int c;
 	long long int d;
 	float f;
-
-printf("size of a char: %lu\n", sizeof(a));
-printf("size of a int: %lu\n", sizeof(b));
-printf("size of a long int: %lu\n", sizeof(c));
-printf("size of a long long int: %lu\n", sizeof(d));
-printf("size of a float: %lu\n", sizeof(f));
+printf("size of a char: %d\n", (unsigned long)sizeof(a));
+printf("size of a int: %d\n", (unsigned long)sizeof(b));
+printf("size of a long int: %d\n", (unsigned long)sizeof(c));
+printf("size of a long long int: %d\n", (unsigned long)sizeof(d));
+printf("size of a float: %d\n", (unsigned long)sizeof(f));
 return (0);
 }
